+++ package.json
@@ -26,9 +26,11 @@
     "@aresrpg/aresrpg-protocol": "^1.3.0",
     "@mongodb-js/zstd": "^1.2.0",
     "alea": "^1.0.1",
+    "aresrpg-engine": "file:../aresrpg-engine",
     "ndarray": "^1.0.19",
     "simplex-noise": "^4.0.1",
-    "sparse-octree": "^7.1.8"
+    "sparse-octree": "^7.1.8",
+    "three": "^0.162.0"
   },
   "devDependencies": {
     "@babel/core": "^7.23.9",
